requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Database
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5

# Authentication
Flask-Login==0.6.3

# Forms and CSRF
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# Development and Testing
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.2.0

# Production
gunicorn==21.2.0
